refactor: piece_initializer 패키지 제거
- piece 패키지로 통합
