+++ src/test/java/BoardTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import piece.Piece;
 import piece.PieceType;
-import piece_initiaizer.StaticPieceInitializer;
+import piece.StaticPieceInitializer;
 import position.Position;
 import position.PositionFile;
 import testUtil.TestConstant;
